+++ web/CHANGES.md
@@ -3,6 +3,10 @@
 - remove generic type param from `IntoCtx` trait
 - bump MSRV to `1.78`
 - update `xitca-http` to `0.6.0`
+- update `xitca-server` to `0.4.0`
+- update `xitca-service` to `0.2.0`
+- update `xitca-tls` to `0.4.0`
+- update `xitca-unsafe-collection` to `0.2.0`
 
 # 0.5.0
 ## Add