+++ web/CHANGES.md
@@ -1,6 +1,12 @@
-# unreleased 0.5.1
+# unreleased 0.6.0
 ## Change
 - remove generic type param from `IntoCtx` trait
+- bump MSRV to `1.79`
+- update `xitca-http` to `0.6.0`
+- update `xitca-server` to `0.4.0`
+- update `xitca-service` to `0.2.0`
+- update `xitca-tls` to `0.4.0`
+- update `xitca-unsafe-collection` to `0.2.0`
 
 # 0.5.0
 ## Add